Refactoring & Transform fix
View matrix now based on inverse of world

+++ Source/GameObject.cpp
@@ -7,7 +7,7 @@ GameObject::GameObject()
 
 GameObject::~GameObject()
 {
-	for (int i = 0; i < components.size(); ++i)
+	for (size_t i = 0; i < components.size(); ++i)
 	{
 		delete components[i];
 	}
